Add `almostFullScreen` command
This change adds a command that used to be in the Spectacle app, to
resize a window to take up most but not all of the screen size. It is
bound to the same keybinding as `fullScreen`, but adding the Shift key.

The resized window is then centered, leaving some padding space
between the window and the borders of the desktop.